requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Google Inline/app.py: 9,11
beautifulsoup4 == 4.6.0

# Google Inline/app.py: 16
lxml == 4.5.2

# Google Inline/app.py: 2,3,6
python_telegram_bot == 12.8

# Google Inline/app.py: 7,10
requests == 2.22.0

# Google Inline/app.py: 13
six == 1.14.0

tinydb
pymongo
analytics-python